+++ src/components/SignupPage/SignupPage.view.tsx
@@ -1,38 +1,15 @@
 import React from 'react';
-import { Controller, UseFormReturn } from 'react-hook-form';
-
-import { Select } from 'chakra-react-select';
-
 import {
   Box,
-  BoxProps,
-  Button,
-  ChakraProps,
-  CheckboxGroup,
-  FormLabelProps,
-  Input,
-  InputProps,
-  Radio,
-  RadioGroup,
-  Stack,
-  Checkbox,
   Text
 } from '@chakra-ui/react';
-
-import FormHelper from '@components/common/FormHelper';
-
-import { FormDataType } from './_hooks/useExampleForm';
-import CheckLineIcon from '@components/common/New/@Icons/System/CheckLine';
-import CheckCircleIcon from '@components/common/New/@Icons/System/CheckCircle';
 import PrimaryButton from '@components/common/New/PrimaryButton';
-import { FormLabelProp, FormPageProps } from './SignupPage.type';
+import { SignupPageProps } from './SignupPage.type';
 import ProfileForm from './_fragments/ProfileForm';
 import AdditionalInfoForm from './_fragments/AdditionalInfoForm';
 import TermsOfServiceForm from './_fragments/TermsOfServiceForm';
 
-
-
-const FormPageView = ({
+const SignupPageView = ({
   formData: {
     register,
     control,
@@ -41,10 +18,7 @@ const FormPageView = ({
   },
   onSubmit,
   ...basisProps
-}: FormPageProps) => {
-  const [checkedItems, setCheckedItems] = React.useState([false, false, false])
-  const allChecked = checkedItems.every(Boolean)
-  const isIndeterminate = checkedItems.some(Boolean) && !allChecked
+}: SignupPageProps) => {
   console.log("errors:", errors)
   return (
     <Box as="form" onSubmit={onSubmit} px="16px" {...basisProps}>
@@ -57,4 +31,4 @@ const FormPageView = ({
   );
 };
 
-export default FormPageView;
+export default SignupPageView;
